backend/ai/OCR_EVoting_Project/requirements.txt
tensorflow
numpy
opencv-python
Pillow
scikit-learn
